functions/requirements.txt
firebase-admin>=6.5.0
firebase-functions>=0.4.2
google-genai>=1.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.13
google-cloud-firestore>=2.19.0
Pillow>=10.0.0
pinecone>=3.0.0
